requirements.txt
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.6
base58==2.1.0
bitarray==1.2.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cytoolz==0.11.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.5.0
django-ratelimit==3.0.1
djangorestframework==3.11.1
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
google-api-core==1.26.3
google-auth==1.29.0
google-cloud-core==1.6.0
google-cloud-logging==2.3.1
googleapis-common-protos==1.53.0
grpcio==1.37.0
gunicorn==20.1.0
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
netaddr==0.8.0
oauthlib==3.1.0
packaging==20.9
parsimonious==0.8.1
proto-plus==1.18.1
protobuf==3.15.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycryptodome==3.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rlp==2.0.1
rsa==4.7.2
sentry-sdk==1.0.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
toolz==0.11.1
urllib3==1.26.4
varint==1.0.2
web3==5.18.0
websockets==8.1
whitenoise==5.2.0
